requirements.txt
asgiref==3.5.2
backports.zoneinfo;python_version<"3.9"
beautifulsoup4==4.11.1
certifi==2022.9.24
charset-normalizer==2.1.1
contourpy==1.0.5
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.15
django-bootstrap4==22.2
django-crispy-forms==1.14.0
django-import-export==2.9.0
django-rangefilter==0.9.1
django-relativedelta==1.1.2
et-xmlfile==1.1.0
fonttools==4.37.3
gunicorn==20.1.0
idna==3.4
kiwisolver==1.4.4
MarkupPy==1.14
matplotlib==3.6.0
numpy==1.23.3
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tablib==3.2.1
urllib3==1.26.12
whitenoise==6.2.0
xlrd==2.0.1
xlwt==1.3.0
